fix(manage-validator): move into separate page folder
