+++ src/pages/ManageValidator/Footer.tsx
@@ -2,7 +2,7 @@ import { Text as TextStat } from '@status-im/components'
 
 import { Text, XStack } from 'tamagui'
 import { ExternalIcon } from '@status-im/icons'
-import Icon from '../../../components/General/Icon'
+import Icon from '../../components/General/Icon'
 const Footer = () => {
   return (
     <XStack
