+++ src/pages/ManageValidator/RightManageSettings/Panels/PanelWrapper.tsx
