+++ src/pages/ManageValidator/ValidatorBalance.tsx
@@ -1,6 +1,6 @@
 import { Text } from '@status-im/components'
 import { Stack, XStack, YStack } from 'tamagui'
-import Icon from '../../../components/General/Icon'
+import Icon from '../../components/General/Icon'
 
 const ValidatorBalance = () => {
   return (
