+++ src/pages/ManageValidator/ValidatorDataCards.tsx
