+++ src/pages/ManageValidator/ValidatorGraffiti.tsx
